Feat: Enforce pawn movement rules
Pawns can now only advance in a straight line, 1 square at a time, or into a
diagonal square occupied by an enemy piece. They also cannot move into or over
a square occupied by an allied piece.

[Pawn's first movement detection on Game#move_piece_to]
The change of state of pawns' @start_position was implemented on #move_piece_to
because when the movement is completed we can be assured that the pawn isn't on
it's start position anymore.

--- The move restrictions probaby still don't work close to the board limits
--- Sometimes an illegal move will crash the program instead of prompting for
another target.