Feat: Execute the castling move
If the castling moves were implemented as simple valid moves for the king the
king's leap would be executed in #move_piece_to, but the tower wouldn't move
simultaneously. It would have no way of knowing that besides moving the king it
was supposed to also move the tower. Thus the implementation of the castling
moves check on Game#move_check? so that a castling move choice can be detected
as such before the #move_piece_to execution and the tower can be moved.

King and Tower instances now rely on the start position flag attribute just
the pawns. Game#move_piece_to now changes the start position attribute of
any piece that responds to the #start_position method.